+++ apps/sample/tailwind.config.ts
@@ -1,11 +1,12 @@
+import type { Config } from 'tailwindcss'
 import { dynamicIconsPlugin } from '@egoist/tailwindcss-icons'
 import { srcubeUI } from '@srcube-taro/theme'
+import animated from 'tailwindcss-animated'
 
-/** @type {import('tailwindcss').Config} */
 export default {
   content: [
     './src/**/*.{js,jsx,ts,tsx}',
-    // '../../node_modules/@srcube-taro/theme/dist/**/*.{js,jsx,ts,tsx}',
+    '../../node_modules/@srcube-taro/theme/dist/**/*.{js,jsx,ts,tsx}',
     '../../packages/core/theme/src/**/*.{js,jsx,ts,tsx}',
   ],
   theme: {
@@ -16,5 +17,6 @@ export default {
   plugins: [
     srcubeUI(),
     dynamicIconsPlugin(),
+    animated,
   ],
-}
+} satisfies Config