+++ packages/core/theme/src/plugin/index.ts
@@ -42,6 +42,30 @@ export function srcubeUI() {
               warning: withDefaultColor('amber'),
               danger: withDefaultColor('rose'),
             },
+            keyframes: {
+              'modal-in': {
+                '0%': { opacity: '0', transform: 'scale3d(1.05, 1.05, 1.05)' },
+                '100%': { opacity: '1', transform: 'scale3d(1, 1, 1)' },
+              },
+              'modal-out': {
+                '0%': { opacity: '1', transform: 'scale3d(1, 1, 1)' },
+                '100%': { opacity: '0', transform: 'scale3d(1.05, 1.05, 1.05)' },
+              },
+              'fade-in': {
+                '0%': { opacity: '0' },
+                '100%': { opacity: '1' },
+              },
+              'fade-out': {
+                '0%': { opacity: '1' },
+                '100%': { opacity: '0' },
+              },
+            },
+            animation: {
+              'modal-in': 'modal-in var(--default-animation-duration, 0.5s) var(--default-animation-timing-function, ease) var(--default-animation-dely, 0s) both',
+              'modal-out': 'modal-out var(--default-animation-duration, 0.5s) var(--default-animation-timing-function, ease) var(--default-animation-dely, 0s) both',
+              'fade-in': 'fade-in var(--default-animation-duration, 0.5s) var(--default-animation-timing-function, ease) var(--default-animation-dely, 0s) both',
+              'fade-out': 'fade-out var(--default-animation-duration, 0.5s) var(--default-animation-timing-function, ease) var(--default-animation-dely, 0s) both',
+            },
           },
         },
       }