+++ packages/core/ui/CHANGELOG.md
@@ -1,10 +1,24 @@
 # @srcube-taro/ui
 
+## 0.0.4
+
+### Patch Changes
+
+- Components and hooks package
+  modal
+- Updated dependencies
+  - @srcube-taro/theme@0.0.4
+  - @srcube-taro/modal@0.0.1
+  - @srcube-taro/button@0.0.4
+  - @srcube-taro/input@0.0.2
+  - @srcube-taro/layouts@0.0.4
+  - @srcube-taro/spinner@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes
 
-- Components:
+- Components
   input
 - Updated dependencies
   - @srcube-taro/theme@0.0.3